+++ src/pages/User/Menu/Menu.jsx
@@ -20,7 +20,7 @@ class Menu extends React.Component {
 
   render() {
     const { isMenuVisisble, dispatch } = this.props;
-    const { screenshotsFound, screenshotsAdded } = this.props.userData;
+    const { solvedScreenshots, addedScreenshots } = this.props.userData;
     return (
       <div className={`Menu ${isMenuVisisble ? '-visisble' : '-hidden'}`}>
         <button
@@ -53,7 +53,7 @@ class Menu extends React.Component {
           </svg>
           Moderation
         </MenuItem> */}
-        <MenuItem to="/user/found" dispatch={dispatch}>
+        <MenuItem to="/user/solved" dispatch={dispatch}>
           <svg
             className="Menu_item_icon"
             xmlns="http://www.w3.org/2000/svg"
@@ -63,7 +63,7 @@ class Menu extends React.Component {
           >
             <path d="M12 0c-6.627 0-12 5.373-12 12s5.373 12 12 12 12-5.373 12-12-5.373-12-12-12zm-1.25 17.292l-4.5-4.364 1.857-1.858 2.643 2.506 5.643-5.784 1.857 1.857-7.5 7.643z" />
           </svg>{' '}
-          {screenshotsFound} screenshots solved
+          {solvedScreenshots} Solved Screenshots
         </MenuItem>
         <MenuItem to="/user/added" dispatch={dispatch}>
           <svg
@@ -75,7 +75,7 @@ class Menu extends React.Component {
           >
             <path d="M12 0c-6.627 0-12 5.373-12 12s5.373 12 12 12 12-5.373 12-12-5.373-12-12-12zm6 13h-5v5h-2v-5h-5v-2h5v-5h2v5h5v2z" />
           </svg>
-          {screenshotsAdded} screenshots added
+          {addedScreenshots} Added Screenshots
         </MenuItem>
       </div>
     );